Moved authserver into a shared project.
