+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: AuthServer
+name: ServerStack
 description: A sample command-line application
 dependencies:
   args: any
